backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
torch==2.2.0
torchvision==0.17.0
timm==0.9.16
grad-cam==1.5.0
opencv-python-headless==4.9.0.80
Pillow==10.2.0
numpy==1.26.4
matplotlib==3.8.3
scikit-learn==1.4.1.post1
fpdf2==2.7.9